Fixed InstructionsSensorCount bug
Previous API allowed Packets that had fewer entries than sensors that
SensorData stored. The original goal of this was to allow old packets to
continue to work seamlessly if a new sensor was added to the end of the
SensorInfo tuple. This was a misjudgement in API design and would
actually produce an obscure error if this were to occur. This has been
rectified. All functions now expect a packet that has an entry for each
sensor that SensorData stores.